Abstract out the generator interface
